requirements.txt
google-cloud-pubsub==2.29.0
